PARTIAL -- fix bug in selectors and improve module naming
