+++ _includes/js/renderers/instagram-modal.js
@@ -17,39 +17,35 @@ export default ({dom, photo}) => {
   const template = dom.select('#instagram-modal-template').content;
   const content = template.cloneNode(true);
 
-  const {
-    images: {
-      standard_resolution: {
-        url
-      } = {}
-    } = {}
-  } = photo;
-
   content.querySelector('.ig-comments-count').textContent = photo.commentsCount;
   content.querySelector('.ig-likes-count').textContent = photo.likesCount;
-  content.querySelector('.ig-modal-image').src = url || '';
+  content.querySelector('.img-modal-avatar').src = photo.profilePicture;
+  content.querySelector('.ig-modal-image').src = photo.images.standard_resolution.url;
   content.querySelector('.ig-modal-link').href = photo.link;
   content.querySelector('.ig-modal-location').textContent = photo.locationName;
   content.querySelector('.ig-modal-text').textContent = photo.text;
   content.querySelector('.ig-time-ago').textContent = getRelativeTimeSinceString(photo.createdAt);
 
-  const closeModal = () => {
-    console.log('closing modal...');
+  const usernameLink = content.querySelector('.ig-modal-username-link');
+  usernameLink.href = `https://instagram.com/${photo.username}`;
+
+  const closeModal = event => {
     dom.select('#instagram-overlay').remove();
     dom.select('body').classList.remove('ig-modal-open');
+    event.preventDefault();
   };
-
-  const node = document.importNode(content, true);
-  node.querySelector('.ig-modal-close').addEventListener('click', closeModal);
-  dom.select('body').classList.add('ig-modal-open');
 
   const handleCloseClick = event => {
     if (event.key === 'Escape') {
-      dom.select('#instagram-overlay').remove();
-      dom.select('body').classList.remove('ig-modal-open');
+      closeModal(event);
       event.target.removeEventListener('keyup', handleCloseClick);
     }
   };
+
+  const node = document.importNode(content, true);
+  dom.select('body').classList.add('ig-modal-open');
+
+  node.querySelector('.ig-modal-close').addEventListener('click', closeModal);
   document.addEventListener('keyup', handleCloseClick);
 
   container.appendChild(node);
