+++ _includes/template/instagram/modal.html
@@ -14,15 +14,16 @@
 
         <div class="modal-head margin-b-1">
           <div class="avatar">
-            <img class="avatar-circle" src="https://scontent.cdninstagram.com/vp/9a1adcbfa769f272e0c36432df814167/5D0724C6/t51.2885-19/s150x150/44209755_979982635518910_5496386888413478912_n.jpg?_nc_ht=scontent.cdninstagram.com"
-              width="40" />
+            <img
+              class="img-modal-avatar avatar-circle"
+              src=""
+              width="40"
+            />
           </div>
           <div class="head-content">
-            <div id="ig-username">
-              <a href="https://instagram.com/c1v0" title="c1v0 on Instagram">
-                <strong class="ig-modal-username">c1v0</strong>
-              </a>
-            </div>
+            <a class="ig-modal-username-link" href="" title="">
+              <strong class="ig-modal-username">c1v0</strong>
+            </a>
             <div class="ig-modal-location txt-small txt-muted"></div>
           </div>
         </div>
