+++ assets/styles/components/_photos.scss
@@ -19,7 +19,7 @@ body.ig-modal-open {
 }
 
 #instagram-overlay {
-    background: rgba(0,0,0,0.5);
+    background: rgba(0,0,0,0.6);
     position: fixed;
     color: white;
     z-index: 104;
